visual-search-app/backend/requirements.txt
# ─── Core API Framework ───────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
aiofiles==23.2.1

# ─── ML / Embeddings ──────────────────────────────────────────────────────────
sentence-transformers==3.0.1
torch==2.3.1
torchvision==0.18.1
Pillow==10.3.0

# ─── Vector Search ────────────────────────────────────────────────────────────
faiss-cpu==1.8.0
numpy==1.26.4

# ─── Metadata (styles.csv preferred — styles.xlsx needs these) ────────────────
pandas==2.2.2
openpyxl==3.1.3

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
